Adds unit selection to ingredient modal
Enables users to select a unit of measurement (e.g., g, ml, piece) when adding or editing ingredients.

This change introduces a new "unit" field to the ingredient type and updates the ingredient modal to include a unit selection dropdown.  It also adjusts the display of nutritional information to be "per unit" instead of "per 100g," providing more flexibility and accuracy in recipe creation.

The broader area selection in create recipe page is also updated for better UX.